add metrics provider
